The end

+++ application/controllers/careers.php
@@ -16,6 +16,7 @@ class Careers extends CI_Controller {
 		$this->load->library('validationsLib');//loading a model_usuario, name of instance
 		/*$this->form_validation->set_message('required','Digite %s');*/
 		$this->form_validation->set_message('careersOk','Este código ya existe');
+		$this->form_validation->set_message('careersOk','Esta carrera ya contiene cursos, no se puede borrar');
 	}
 
 	public function index()
@@ -43,8 +44,18 @@ class Careers extends CI_Controller {
 	{
 		$register = $this->input->post();
 
-		$this->form_validation->set_rules('code', 'Código', 'required');
-		$this->form_validation->set_rules('name', 'Nombre', 'required');
+		//if are same
+		if($this->input->post('code')===$this->input->post('code1'))
+		{
+			$this->form_validation->set_rules('code', 'Código', 'required');
+			$this->form_validation->set_rules('name', 'Nombre', 'required');
+		}
+		//if the new identification_card is changing(two varibles are different)
+		else
+		{
+			$this->form_validation->set_rules('code', 'Código', 'required|callback_careersOk');
+			$this->form_validation->set_rules('name', 'Nombre', 'required');
+		}
 
 		//if it is false is because any of the rules over is failing
 		if ($this->form_validation->run()==FALSE) {
@@ -52,6 +63,7 @@ class Careers extends CI_Controller {
 		}
 		else
 		{
+			unset($register['code1']);//delete the last identification_card
 			$register['updated']=date('Y/m/d H:i');
 			$this->m->update($register);
 			redirect('careers/index');
@@ -96,13 +108,34 @@ class Careers extends CI_Controller {
 		}
 		
 	}
+	public function errorErase($id)
+	{
+		return $this->validationslib->careersDelete($id);
+	}
+	public function error()
+	{
+		return FALSE;
+	}
 
 
 	public function delete($id)
 	{
-		$this->m->delete($id);
-		redirect("careers/index");
- 
+
+		//var_dump($this->input->post());
+
+		
+		$status=$this->errorErase($id);
+		//if it is existing in courses
+		if(!$status)
+		{
+			print "<script type=\"text/javascript\">alert('Esta carrera contiene cursos, no se puede borrar');</script>";
+			$this->edit($id);//going to ingreso() without lossing the values of the variables 
+		}
+		else
+		{
+			$this->m->delete($id);
+			redirect("careers/index");
+ 		}
 	}
 
 }
